Conflicts:
	cgogn/core/cmap/cmap2.h

+++ cgogn/geometry/types/plane_3d.h
@@ -141,8 +141,8 @@ private:
 #if defined(CGOGN_USE_EXTERNAL_TEMPLATES) && (!defined(GEOMETRY_TYPES_PLANE_3D_CPP_))
 extern template class CGOGN_GEOMETRY_API Plane3D<Eigen::Vector3d>;
 extern template class CGOGN_GEOMETRY_API Plane3D<Eigen::Vector3f>;
-extern template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<double,3>>>;
-extern template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float,3>>>;
+extern template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float32,3>>>;
+extern template class CGOGN_GEOMETRY_API Plane3D<Vec_T<std::array<float64,3>>>;
 #endif // defined(CGOGN_USE_EXTERNAL_TEMPLATES) && (!defined(GEOMETRY_TYPES_PLANE_3D_CPP_))
 
 } // namespace geometry
